pyproject.toml
[tool.poetry]
name = "google-ads-api"
version = "0.1.0"
description = "API Client for Google Ads API"
authors = ["bradi <bradisbon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.1"
pyyaml = "^6.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
